feat((documents)): added Http route functions to documents service and completed mock server
Completed a mock server to simulate the data dynamics of the case studies. Added document management
functions and api routes to the documentsService. Slight modification to interfaces to match the
objects expected by the server. Changed document view to a full paged table with alert to create a
new case study.