feat(integration): update name conventions, integrated date procesing, separated models from mock
Updated the models to match the responses from the server and to match the expected structure of the
request body. Separated the data structures used on the mock server from those used on the web app.
Added maximum length validator to fields on the create document form. Added date processing to the
case document resolver, now the dates can come from the server as the string "YYYY-MM-DD" and be
stored on the web app as date object neutral to the timezone. Added language fieldto the case
document creation form. Initial Documentation of parts of the authentification module.