feat(dashboard): Added actions column to table and modified its data handler to match server
The table used to fetch the data on the server directly as json wich worked on the development environment but not with the server. Now the get documents request is handled by the main document dashboard component, wish upon sucessful retrieval of the information will send the data to the table.Several test api routes on the services were renamed to match the current test environment.

doctable api was changed