feat(integration): Fixed routes for section editor and simple parameters
The server rejected some of the section routes due to a disparity of the body of the requests. Also the sections in the ui were adressed from 0, whereas in the server they are addressed from 1. Changed the addresing to start from 1 for sections. Section 0 now shall redirect to invalid page.

Known errors are actors, authors, timeline and options from general routes.