feat(error handling): added Refresh Token Interceptors and editor error notifications
Added requestErrorInterceptor to intercept 401 errors from all the requests on the app. Upon
catching the authorization errors, a refresh of the token is tried to gain a new valid access token.
If failed after refresh, a new expire method from the authentication service will clear the invalid
session and the user shall be redirected to the login page with a session expired message. This
message will ocurr also the first time a user enters the system after not succesfully loggin out,
since there is still session data on his computer. This interceptor was implemented to take into
consideration concurrent 401 errors with the help of a subject and flag to disable concurrent
refresh requests. Document edition service was modified to account for error messaging. However, the
user still needs to click edit again to fix the errors.