feat(editor): added canDeactivate guard for the section editor and fixed sticky toolbar
Toolbar was not sticking to the top when writing long documents, now it sticks. User was able to
change sections without warning, an issue that lead to potential data loss. A canDeactivate guard
based on the section component already implemented is saved flag was developed to fire a warning to
the user if the automated saved of section has not been completed yet. Also a loading semaphore was
added, such to fix an issue that the first load of a section always lead to a un-needed save request
due to a modification of the editor contents due to load. This semaphore is now incremented when
loading and decremented when finishing loading, such that the semafore reaches a falsy state 0 when
the last section on a series of fast changes is made.