fix(dashboard): fixed author and actor validators and style on document creation form
Author and Actors on the document creation form had a validator error that permited to submit the
form when the any actor or author that was not the original card present is still invalid. This was
fixed by eliminating the control access on the component file to the form array. An error message
for when the email is not from the upr domain was added. Style was adjusted on the doc creation form
to have smaller buttons and more spacing. Added a close button to the doc creationdialog, disabled
click outside closing.