fix(dashboard): deactivated create case study button on loading
An active create case study button on loading created the chance that an extra case study be created
after reaching the maximum limit allowed, since the amount of case studies a collaborator had is
unkown at the loading state. A conditional was added to the button to show the button only when the
page is not loading.